Use external webbrowser.Open
