+++ bakery/example/main.go
@@ -22,6 +22,7 @@ import (
 	"log"
 	"net"
 	"net/http"
+	"net/url"
 
 	"gopkg.in/macaroon-bakery.v1/bakery"
 	"gopkg.in/macaroon-bakery.v1/httpbakery"
@@ -68,4 +69,15 @@ func serve(newHandler func(string) (http.Handler, error)) (endpointURL string, e
 	}
 	go http.Serve(listener, handler)
 	return endpointURL, nil
+}
+
+func newClient() *httpbakery.Client {
+	return &httpbakery.Client{
+		Client: httpbakery.NewHTTPClient(),
+		VisitWebPage: func(url *url.URL) error {
+			fmt.Printf("please visit this web page:\n")
+			fmt.Printf("\t%s\n", url)
+			return nil
+		},
+	}
 }
